requirements.txt
Flask==2.3.2
requests==2.31.0
pytest==7.4.2
pytest-dependency==0.5.1
allure-pytest==2.14.3
